Add junit-host and make junit an alias for it
The junit build target breaks naming conventions. A target built
for the host should have a -host suffix. This change is the
first step in correcting that before adding any new targets.

The next step is to replace all usages of junit target to
junit-host, including some in goog/master.

The step after that is to remove the junit target itself.

Bug: 30188076
Test: make checkbuild
Change-Id: Icaa8770a460b06e514ecc8813f6d70c2b9c526d3