Minor changes to JUnit 4.10 to make it more like 4.12
Some other projects break when built against 4.12. This allows
those projects to be updated separately to the JUnit 4.12
upgrade.

Assert.java and Description.java are copied directly from
upstream 4.12; that brings in new methods from 4.12. The changes
are so extensive because the upstream code was reformatted
between 4.10 anbd 4.12.

Between 4.10 and 4.12 RuleFieldValidator.java was renamed to
RuleMemberValidator.java. To allow existing code that uses
RuleFieldValidator to be modified to use RuleMemberValidator
instead RuleFieldValidator.java has been copied to
RuleMemberValidator.java.

The change to BlockJUnit4ClassRunner reflects an upstream change
made between 4.10 and 4.12.

Bug: 33613916
Test: make checkbuild
Change-Id: Id8ee6d08268152ec1a22a309c4582f8f6663ae88